+++ tests/ObjectResolverTest.php
@@ -2,7 +2,7 @@
 
 namespace Nandan108\PropAccess\Tests;
 
-use Nandan108\PropAccess\AccessorRegistry;
+use Nandan108\PropAccess\PropAccess;
 use Nandan108\PropAccess\Tests\Fixtures\SampleEntity;
 use PHPUnit\Framework\TestCase;
 
@@ -12,9 +12,9 @@ final class ObjectResolverTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
-        AccessorRegistry::bootDefaultResolvers();
+        PropAccess::bootDefaultResolvers();
         // A second time for coverage of the already-booted case
-        AccessorRegistry::bootDefaultResolvers();
+        PropAccess::bootDefaultResolvers();
     }
 
     public function testObjectGetterResolver(): void
@@ -22,7 +22,7 @@ final class ObjectResolverTest extends TestCase
         $entity = new SampleEntity();
 
         /** @var array<string, \Closure(mixed): mixed> $map */
-        $map = AccessorRegistry::getGetterMap($entity);
+        $map = PropAccess::getGetterMap($entity);
 
         $this->assertArrayHasKey('plain', $map);
         $this->assertArrayHasKey('hidden', $map);
@@ -36,9 +36,12 @@ final class ObjectResolverTest extends TestCase
 
         // however, a direct request for the public property should still work
         /** @var array<string, \Closure(mixed): mixed> $directRequestMap */
-        $directRequestMap = AccessorRegistry::getGetterMap($entity, ['public_snake_case']);
+        $directRequestMap = PropAccess::getGetterMap($entity, ['public_snake_case']);
         $this->assertArrayHasKey('public_snake_case', $directRequestMap);
         $this->assertSame('snake', $directRequestMap['public_snake_case']($entity));
+
+        $directRequestMapForSingleProp = PropAccess::getGetterMap($entity, 'public_snake_case');
+        $this->assertSame($directRequestMap, $directRequestMapForSingleProp);
 
         // should have key because getter is public
         $this->assertArrayHasKey('privateSnakeCase', $map);
@@ -56,7 +59,7 @@ final class ObjectResolverTest extends TestCase
         $this->expectExceptionMessage('No public getter or property found for: nonExisting, missingProperty in Nandan108\PropAccess\Tests\Fixtures\SampleEntity');
 
         $entity = new SampleEntity();
-        AccessorRegistry::getGetterMap($entity, ['plain', 'nonExisting', 'missingProperty'], false);
+        PropAccess::getGetterMap($entity, ['plain', 'nonExisting', 'missingProperty'], false);
     }
 
     public function testObjectSetterResolver(): void
@@ -64,7 +67,7 @@ final class ObjectResolverTest extends TestCase
         $entity = new SampleEntity();
 
         /** @var array<string, \Closure(mixed, mixed): void> $setterMap */
-        $setterMap = AccessorRegistry::getSetterMap($entity);
+        $setterMap = PropAccess::getSetterMap($entity);
 
         $this->assertArrayHasKey('plain', $setterMap);
         $this->assertArrayHasKey('hidden', $setterMap);
@@ -79,7 +82,7 @@ final class ObjectResolverTest extends TestCase
 
         // however, a direct request for the public property should still work
         /** @var array<string, \Closure(mixed, mixed): void> $directRequestMap */
-        $directRequestMap = AccessorRegistry::getSetterMap($entity, ['public_snake_case']);
+        $directRequestMap = PropAccess::getSetterMap($entity, ['public_snake_case']);
         $this->assertArrayHasKey('public_snake_case', $directRequestMap);
         $directRequestMap['public_snake_case']($entity, 'SNAKE');
         /** @psalm-suppress DocblockTypeContradiction */
@@ -105,6 +108,6 @@ final class ObjectResolverTest extends TestCase
         $this->expectExceptionMessage('No public setter or property found for: nonExisting, missingProperty in Nandan108\PropAccess\Tests\Fixtures\SampleEntity');
 
         $entity = new SampleEntity();
-        AccessorRegistry::getSetterMap($entity, ['plain', 'nonExisting', 'missingProperty'], false);
+        PropAccess::getSetterMap($entity, ['plain', 'nonExisting', 'missingProperty'], false);
     }
 }