+++ tests/PropAccessTest.php
@@ -2,40 +2,41 @@
 
 namespace Nandan108\PropAccess\Tests;
 
-use Nandan108\PropAccess\AccessorRegistry;
+use Nandan108\PropAccess\PropAccess;
 use Nandan108\PropAccess\Tests\Fixtures\SampleEntity;
 use PHPUnit\Framework\TestCase;
 
-final class AccessorRegistryTest extends TestCase
+final class PropAccessTest extends TestCase
 {
     #[\Override]
     protected function setUp(): void
     {
         parent::setUp();
-        AccessorRegistry::bootDefaultResolvers();
+        PropAccess::bootDefaultResolvers();
     }
 
     public function testNoGetterResolverAvailable(): void
     {
         $this->expectException(\LogicException::class);
-        $this->expectExceptionMessage('No getter resolver supports type "string"');
+        $this->expectExceptionMessage('No getter resolver supports type "SplObjectStorage"');
 
-        AccessorRegistry::getGetterMap('not an object');
+        /** @psalm-suppress InvalidArgument */
+        PropAccess::getGetterMap(new \SplObjectStorage());
     }
 
     public function testNoSetterResolverAvailable(): void
     {
-        $this->assertNull(AccessorRegistry::getSetterMap('not an object', throwOnNotFound: false));
+        $this->assertNull(PropAccess::getSetterMap('not an object', throwOnNotFound: false));
         $this->expectException(\LogicException::class);
         $this->expectExceptionMessage('No setter resolver supports type "string"');
-        AccessorRegistry::getSetterMap('not an object');
+        PropAccess::getSetterMap('not an object');
     }
 
     // canGetGetterMap() for a resolable type
     public function testCanGetGetterMapForResolvableType(): void
     {
         $source = new SampleEntity();
-        $canGetMap = AccessorRegistry::canGetGetterMap($source);
+        $canGetMap = PropAccess::canGetGetterMap($source);
         $this->assertTrue($canGetMap, 'Expected to be able to get getter map for SampleEntity');
     }
 
@@ -43,7 +44,7 @@ final class AccessorRegistryTest extends TestCase
     public function testCanGetGetterMapForNonResolvableType(): void
     {
         $source = new \SplObjectStorage();
-        $canGetMap = AccessorRegistry::canGetGetterMap($source);
+        $canGetMap = PropAccess::canGetGetterMap($source);
         $this->assertFalse($canGetMap, 'Expected not to be able to get getter map for SplObjectStorage');
     }
 
@@ -51,7 +52,7 @@ final class AccessorRegistryTest extends TestCase
     public function testCanGetSetterMapForResolvableType(): void
     {
         $target = new SampleEntity();
-        $canGetMap = AccessorRegistry::canGetSetterMap($target);
+        $canGetMap = PropAccess::canGetSetterMap($target);
         $this->assertTrue($canGetMap, 'Expected to be able to get setter map for SampleEntity');
     }
 
@@ -59,7 +60,7 @@ final class AccessorRegistryTest extends TestCase
     public function testCanGetSetterMapForNonResolvableType(): void
     {
         $target = new \SplObjectStorage();
-        $canGetMap = AccessorRegistry::canGetSetterMap($target);
+        $canGetMap = PropAccess::canGetSetterMap($target);
         $this->assertFalse($canGetMap, 'Expected not to be able to get setter map for SplObjectStorage');
     }
 
@@ -68,15 +69,15 @@ final class AccessorRegistryTest extends TestCase
     {
         $this->expectException(\LogicException::class);
         $this->expectExceptionMessage('No getter resolver supports type "SplObjectStorage"');
-        $this->assertNull(AccessorRegistry::getGetterMap(new \SplObjectStorage(), throwOnNotFound: false));
-        AccessorRegistry::getGetterMap(new \SplObjectStorage());
+        $this->assertNull(PropAccess::getGetterMap(new \SplObjectStorage(), throwOnNotFound: false));
+        PropAccess::getGetterMap(new \SplObjectStorage());
     }
 
     // getValueMap() for a resolable type
     public function testGetValueMapForResolvableType(): void
     {
         $source = new SampleEntity();
-        $map = AccessorRegistry::getValueMap($source);
+        $map = PropAccess::getValueMap($source);
         $this->assertIsArray($map, 'Expected to get a value map for SampleEntity');
         $this->assertArrayHasKey('plain', $map, 'Expected "plain" property to be in the value map');
         $this->assertSame($source->plain, $map['plain'], 'Expected "plain" value to match the entity value');
@@ -86,9 +87,9 @@ final class AccessorRegistryTest extends TestCase
     public function testResolveValuesWithValidMap(): void
     {
         $source = new SampleEntity();
-        $map = AccessorRegistry::getGetterMap($source);
+        $map = PropAccess::getGetterMap($source);
         $this->assertIsArray($map, 'Expected getter map to be an array');
-        $resolved = AccessorRegistry::resolveValues($map, $source);
+        $resolved = PropAccess::resolveValues($map, $source);
         $this->assertSame($source->plain, $resolved['plain'], 'Expected resolved "plain" value to match the entity value');
     }