+++ tests/StdClassResolverTest.php
@@ -2,7 +2,7 @@
 
 namespace Nandan108\PropAccess\Tests;
 
-use Nandan108\PropAccess\AccessorRegistry;
+use Nandan108\PropAccess\PropAccess;
 use PHPUnit\Framework\TestCase;
 
 final class StdClassResolverTest extends TestCase
@@ -11,7 +11,7 @@ final class StdClassResolverTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
-        AccessorRegistry::bootDefaultResolvers();
+        PropAccess::bootDefaultResolvers();
     }
 
     public function testStdClassGetterResolver(): void
@@ -23,7 +23,7 @@ final class StdClassResolverTest extends TestCase
         ];
 
         /** @var array<string, \Closure(mixed): mixed> $map */
-        $map = AccessorRegistry::getGetterMap($stdObj);
+        $map = PropAccess::getGetterMap($stdObj);
 
         $this->assertArrayHasKey('plain', $map);
         $this->assertArrayHasKey('hidden', $map);
@@ -32,7 +32,7 @@ final class StdClassResolverTest extends TestCase
 
         // however, a direct request for the public property should still work
         /** @var array<string, \Closure(mixed): mixed> $directRequestMap */
-        $directRequestMap = AccessorRegistry::getGetterMap($stdObj, ['public_snake_case']);
+        $directRequestMap = PropAccess::getGetterMap($stdObj, ['public_snake_case']);
         $this->assertArrayHasKey('public_snake_case', $directRequestMap);
         $this->assertSame('snake', $directRequestMap['public_snake_case']($stdObj));
     }
@@ -46,11 +46,12 @@ final class StdClassResolverTest extends TestCase
         ];
 
         /** @var array<string, \Closure(mixed, mixed): void> $map */
-        $map = AccessorRegistry::getSetterMap($stdObj);
+        $map = PropAccess::getSetterMap($stdObj);
         foreach ($map as $key => $setter) {
             $setter($stdObj, "test-$key");
         }
 
+        /** @var array-key $key */
         foreach ($stdObj as $key => $val) {
             $this->assertSame("test-$key", $val);
         }
@@ -63,6 +64,6 @@ final class StdClassResolverTest extends TestCase
         $this->expectExceptionMessage('One or more property not found in \StdClass object: baz, qux');
 
         $entity = (object) ['foo' => 1, 'bar' => 42];
-        AccessorRegistry::getGetterMap($entity, ['foo', 'baz', 'bar', 'qux'], false);
+        PropAccess::getGetterMap($entity, ['foo', 'baz', 'bar', 'qux'], false);
     }
 }