+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [0.4.0] – 2025-06-29
+
+### Changed
+- Renamed `AccessorRegistry` to `PropAccess` for coherence with package name.
+- Renamed `AccessorProxy` to `AccessProxy`.
+- Resolved static analysis issues and bumped Psalm errorLevel to 1
+- Updated README to reflect these changes.
 
 ## [0.3.0] – 2025-06-24